beam/requirements.txt
setuptools
annotated-types==0.7.0
apache-beam[gcp]==2.62.0
attrs==24.3.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudpickle==2.2.1
crcmod==1.7
deprecated==1.2.15
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
docstring-parser==0.16
elastic-transport==8.15.1
elasticsearch==8.15.1
fastavro==1.10.0
fasteners==0.19
google-api-core[grpc]==2.24.0
google-apitools==0.5.31
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.78.0
google-cloud-bigquery==3.29.0
google-cloud-bigquery-storage==2.27.0
google-cloud-bigtable==2.28.1
google-cloud-core==2.4.1
google-cloud-datastore==2.20.2
google-cloud-dlp==3.26.0
google-cloud-language==2.16.0
google-cloud-pubsub==2.27.2
google-cloud-pubsublite==1.11.1
google-cloud-recommendations-ai==0.10.15
google-cloud-resource-manager==1.14.0
google-cloud-secret-manager==2.22.1
google-cloud-spanner==3.51.0
google-cloud-storage==2.19.0
google-cloud-videointelligence==2.15.0
google-cloud-vision==3.9.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos[grpc]==1.66.0
grpc-google-iam-v1==0.14.0
grpc-interceptor==0.15.4
grpcio==1.65.5
grpcio-status==1.65.5
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
importlib-metadata==8.5.0
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.1.0
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
keyrings-google-artifactregistry-auth==1.1.2
lxml==5.3.0
more-itertools==10.6.0
numpy==2.2.2
oauth2client==4.1.3
objsize==0.7.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pydot==1.4.2
pymongo==4.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyyaml==6.0.2
redis==5.2.1
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
shapely==2.0.6
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.3
typing-extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0